moved dialog component files to user-management folder
